+++ app/src/main/res/layout/edit_delete_shape_list_header.xml
@@ -54,6 +54,6 @@
         android:layout_marginLeft="10dp"
         android:layout_weight="0"
         android:visibility="gone"/>
-    android:visibility="gone"
+
 
 </LinearLayout>
