+++ app/src/main/res/layout/fragment_edit_delete_shape.xml
@@ -1,9 +1,9 @@
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout  xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.monash.fit2081.db.EditDeleteShape">
+    tools:context="com.mnov23.Gestures.EditDeleteShape">
 
     <!-- TODO: Update blank fragment layout -->
 
@@ -17,4 +17,4 @@
         app:layout_constraintRight_toRightOf="parent"
         android:layout_marginTop="16dp"
         app:layout_constraintTop_toTopOf="parent"/>
-</android.support.constraint.ConstraintLayout>
+</androidx.constraintlayout.widget.ConstraintLayout >
