+++ app/src/main/res/layout/item_delete_edit.xml
@@ -52,6 +52,5 @@
         android:layout_marginLeft="10dp"
         android:layout_weight="0"
         android:visibility="gone"/>
-    android:visibility="gone"
 
 </LinearLayout>
